requirements.txt
Flask==3.1.0
flask-cors==6.0.0
gunicorn==22.0.0
python-dotenv==1.0.0
